Split events into months for TabNav
